Fix the motherfucking transparency

+++ Object.cxx
@@ -345,7 +345,7 @@ void Object::render(Renderer* renderer) {
     if ( texture ) texture->beginRender(renderer, model);
     model->beginRender(renderer);
 
-    renderer->getCurrentProgram()->setUniform("objectAlpha", this->transparency);
+    renderer->getCurrentProgram()->setUniform("objectAlpha", (float)this->transparency);
 
     model->draw(renderer);
